+++ test/calculator.js
@@ -1,7 +1,5 @@
-'use strict';
-
-var calc = require('../lib/calculator');
-var dataDriven = require('data-driven');
+const calc = require('../src/lib/calculator');
+const dataDriven = require('data-driven');
 
 describe('Calculator testing', function () {
   describe('No resize', function () {
@@ -12,7 +10,7 @@ describe('Calculator testing', function () {
       { x: 100, y: 100, xN: 100, yN: 100, result: { width: 100, height: 100, dx: 0 } },
     ], function() {
       it( '{x}x{y} to {xN}x{yN}', function (p, done) {
-        var params = calc.correctSize(p.x, p.y, p.xN, p.yN);
+        const params = calc.correctSize(p.x, p.y, p.xN, p.yN);
         params.should.eql(p.result);
         done();
       });
@@ -27,7 +25,7 @@ describe('Calculator testing', function () {
       { x: 75,  y: 58,  xN: 99, yN: 99, result: { width: 128, height: 99,  dx: 15 } },
     ], function(a) {
       it( '{x}x{y} to {xN}x{yN}', function (p, done) {
-        var params = calc.correctSize(p.x, p.y, p.xN, p.yN);
+        const params = calc.correctSize(p.x, p.y, p.xN, p.yN);
         params.should.eql(p.result);
         done();
       });