+++ test/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-require('./config');
-require('./calculator');
-require('./modules');