+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # images-manipulation-performance changelog
 
+## [6.0.1](https://github.com/ivanoff/images-manipulation-performance/tree/6.0.1) (2019-06-08)
+[Full Changelog](/compare/5.2.1...6.0.1)
+
+**What Was Done:**
+
+- add new module: image-js
+- add new module: easyimage
+
+
+## [5.2.1](https://github.com/ivanoff/images-manipulation-performance/tree/5.2.1) (2019-06-01)
+[Full Changelog](/compare/1.2.1...5.2.1)
+
+**What Was Done:**
+
+- refactoring
+
+
 ## [1.2.1](https://github.com/ivanoff/images-manipulation-performance/tree/1.2.1) (2016-08-26)
 [Full Changelog](/compare/1.0.1...1.2.1)